+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented here.
 
 ---
 
+## [Unreleased]
+
+### Changed
+
+- tmux_error_handler_assign() - streamlined for performance and added inline comments
+
+---
+
 ## [2.2.3] - 2025-04-28
 
 ### Added
@@ -14,7 +22,7 @@ All notable changes to this project will be documented here.
 
 - Changed titles of some menus to better reflect current placement
 - Removed a relative path expander, no longer needed and caused some external
-menu actions to fail
+  menu actions to fail
 
 ---